+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 
 Latest
 ------
+* Added Checklist and Observation comments - newly available from the eBird API.
 
 0.9.1 (2025-02-03)
 ------------------